went back to int station number to eliminate many casts
git-svn-id: svn://orca.physics.unc.edu/Orca64@11963 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Custom Hardware/SNO+/Fec32/ORFec32Model.m
@@ -1528,7 +1528,7 @@ static int              sChannelsNotChangedCount = 0;
 	return theAveCurrent;
 }
 
-- (NSUInteger) stationNumber
+- (int) stationNumber
 {
 	//we have a weird mapping because fec cards can only be in slots 1-16 and they are mapped to 0 - 15
 	return [[self crate] maxNumberOfObjects] - [self slot] - 2;
